[Style] team 엔티티 이름변경 및 안어울리는 변수 이름 변경, solo,duo,trio Entity 및repository 따로 폴더 생성해줌, 필요없는 주석 삭제.
