+++ src/main/java/com/lolduo/duo/entity/gameInfo/DuoEntity.java
@@ -1,4 +1,4 @@
-package com.lolduo.duo.entity;
+package com.lolduo.duo.entity.gameInfo;
 
 import com.vladmihalcea.hibernate.type.json.JsonStringType;
 import lombok.Getter;
@@ -14,9 +14,9 @@ import java.util.TreeSet;
 @Entity
 @NoArgsConstructor
 @Getter
-@Table(name = "trio")
+@Table(name = "duo")
 @TypeDef(name = "json", typeClass = JsonStringType.class)
-public class TrioEntity {
+public class DuoEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
@@ -26,31 +26,30 @@ public class TrioEntity {
     private Boolean win;
 
     @Type(type = "json")
-    @Column(name = "position")
-    private Map<Long,String> positionMap;
+    @Column(name = "position", columnDefinition = "json")
+    private Map<Long, String> position;
 
     @Type(type = "json")
     @Column(name = "item_list", columnDefinition = "json")
-    private Map<Long,List<Long>> itemListMap;
+    private Map<Long,List<Long>> itemList;
 
     @Type(type = "json")
     @Column(name = "spell_list", columnDefinition = "json")
-    private Map<Long,TreeSet<Long>> spellListMap;
-
+    private Map<Long,TreeSet<Long>> spellList;
     @Type(type = "json")
-    @Column(name = "champion")
-    private TreeSet<Long> championList;
+    @Column(name = "champion",columnDefinition = "json")
+    private TreeSet<Long> champion;
 
     @Type(type = "json")
     @Column(name = "perk_list", columnDefinition = "json")
-    private Map<Long,List<Long>> perkListMap;
+    private Map<Long,List<Long>> perkList;
 
-    public TrioEntity( Boolean win, Map<Long, String> positionMap, Map<Long, List<Long>> itemListMap, Map<Long, TreeSet<Long>> spellListMap, TreeSet<Long> championList, Map<Long, List<Long>> perkListMap) {
+    public DuoEntity(Boolean win, Map<Long, String> position, Map<Long, List<Long>> itemList, Map<Long, TreeSet<Long>> spellList, TreeSet<Long> champion, Map<Long, List<Long>> perkList) {
         this.win = win;
-        this.positionMap = positionMap;
-        this.itemListMap = itemListMap;
-        this.spellListMap = spellListMap;
-        this.championList = championList;
-        this.perkListMap = perkListMap;
+        this.position = position;
+        this.itemList = itemList;
+        this.spellList = spellList;
+        this.champion = champion;
+        this.perkList = perkList;
     }
 }
