+++ src/main/java/com/lolduo/duo/entity/gameInfo/QuintetEntity.java
@@ -1,6 +1,5 @@
-package com.lolduo.duo.entity.solo;
+package com.lolduo.duo.entity.gameInfo;
 
-import com.lolduo.duo.entity.SpellEntity;
 import com.vladmihalcea.hibernate.type.json.JsonStringType;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,14 +8,15 @@ import org.hibernate.annotations.TypeDef;
 
 import javax.persistence.*;
 import java.util.List;
+import java.util.Map;
 import java.util.TreeSet;
 
 @Entity
 @NoArgsConstructor
 @Getter
-@Table(name = "solo")
+@Table(name = "team")
 @TypeDef(name = "json", typeClass = JsonStringType.class)
-public class SoloEntity {
+public class QuintetEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
@@ -25,29 +25,32 @@ public class SoloEntity {
     @Column(name = "win")
     private Boolean win;
 
+    @Type(type = "json")
     @Column(name = "position")
-    private String position;
+    private Map<Long,String> positionMap;
 
     @Type(type = "json")
     @Column(name = "item_list", columnDefinition = "json")
-    private List<Long> itemList;
+    private Map<Long,List<Long>> itemListMap;
 
     @Type(type = "json")
     @Column(name = "spell_list", columnDefinition = "json")
-    private TreeSet<Long> spellList;
+    private Map<Long,TreeSet<Long>> spellListMap;
+
+    @Type(type = "json")
     @Column(name = "champion")
-    private Long champion;
+    private TreeSet<Long> championList;
 
     @Type(type = "json")
     @Column(name = "perk_list", columnDefinition = "json")
-    private List<Long> perkList;
+    private Map<Long,List<Long>> perkListMap;
 
-    public SoloEntity( Boolean win, String position, List<Long> itemList, TreeSet<Long> spellList, Long champion, List<Long> perkList) {
+    public QuintetEntity(Boolean win, Map<Long, String> positionMap, Map<Long, List<Long>> itemListMap, Map<Long, TreeSet<Long>> spellListMap, TreeSet<Long> championList, Map<Long, List<Long>> perkListMap) {
         this.win = win;
-        this.position = position;
-        this.itemList = itemList;
-        this.spellList = spellList;
-        this.champion = champion;
-        this.perkList = perkList;
+        this.positionMap = positionMap;
+        this.itemListMap = itemListMap;
+        this.spellListMap = spellListMap;
+        this.championList = championList;
+        this.perkListMap = perkListMap;
     }
 }
