+++ src/main/java/com/lolduo/duo/entity/gameInfo/SoloEntity.java
@@ -1,5 +1,6 @@
-package com.lolduo.duo.entity;
+package com.lolduo.duo.entity.gameInfo;
 
+import com.lolduo.duo.entity.SpellEntity;
 import com.vladmihalcea.hibernate.type.json.JsonStringType;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,15 +9,14 @@ import org.hibernate.annotations.TypeDef;
 
 import javax.persistence.*;
 import java.util.List;
-import java.util.Map;
 import java.util.TreeSet;
 
 @Entity
 @NoArgsConstructor
 @Getter
-@Table(name = "team")
+@Table(name = "solo")
 @TypeDef(name = "json", typeClass = JsonStringType.class)
-public class TeamEntity {
+public class SoloEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
@@ -25,32 +25,29 @@ public class TeamEntity {
     @Column(name = "win")
     private Boolean win;
 
-    @Type(type = "json")
     @Column(name = "position")
-    private Map<Long,String> positionMap;
+    private String position;
 
     @Type(type = "json")
     @Column(name = "item_list", columnDefinition = "json")
-    private Map<Long,List<Long>> itemListMap;
+    private List<Long> itemList;
 
     @Type(type = "json")
     @Column(name = "spell_list", columnDefinition = "json")
-    private Map<Long,TreeSet<Long>> spellListMap;
-
-    @Type(type = "json")
+    private TreeSet<Long> spellList;
     @Column(name = "champion")
-    private TreeSet<Long> championList;
+    private Long champion;
 
     @Type(type = "json")
     @Column(name = "perk_list", columnDefinition = "json")
-    private Map<Long,List<Long>> perkListMap;
+    private List<Long> perkList;
 
-    public TeamEntity( Boolean win, Map<Long, String> positionMap, Map<Long, List<Long>> itemListMap, Map<Long, TreeSet<Long>> spellListMap, TreeSet<Long> championList, Map<Long, List<Long>> perkListMap) {
+    public SoloEntity( Boolean win, String position, List<Long> itemList, TreeSet<Long> spellList, Long champion, List<Long> perkList) {
         this.win = win;
-        this.positionMap = positionMap;
-        this.itemListMap = itemListMap;
-        this.spellListMap = spellListMap;
-        this.championList = championList;
-        this.perkListMap = perkListMap;
+        this.position = position;
+        this.itemList = itemList;
+        this.spellList = spellList;
+        this.champion = champion;
+        this.perkList = perkList;
     }
 }
