--- src/main/java/com/lolduo/duo/repository/TeamRepository.java
@@ -1,7 +1,0 @@
-package com.lolduo.duo.repository;
-
-import com.lolduo.duo.entity.TeamEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TeamRepository extends JpaRepository<TeamEntity,Long> {
-}
