+++ src/main/java/com/lolduo/duo/repository/gameInfo/DuoRepository.java
@@ -1,6 +1,6 @@
-package com.lolduo.duo.repository;
+package com.lolduo.duo.repository.gameInfo;
 
-import com.lolduo.duo.entity.DuoEntity;
+import com.lolduo.duo.entity.gameInfo.DuoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DuoRepository extends JpaRepository<DuoEntity, Long> {
