+++ src/main/java/com/lolduo/duo/repository/gameInfo/QuintetRepository.java
@@ -1,0 +1,7 @@
+package com.lolduo.duo.repository.gameInfo;
+
+import com.lolduo.duo.entity.gameInfo.QuintetEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface QuintetRepository extends JpaRepository<QuintetEntity,Long> {
+}
