+++ src/main/java/com/lolduo/duo/repository/gameInfo/SoloRepository.java
@@ -1,6 +1,6 @@
-package com.lolduo.duo.repository;
+package com.lolduo.duo.repository.gameInfo;
 
-import com.lolduo.duo.entity.solo.SoloEntity;
+import com.lolduo.duo.entity.gameInfo.SoloEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SoloRepository extends JpaRepository<SoloEntity, Long> {
