+++ src/main/java/com/lolduo/duo/repository/gameInfo/TrioRepository.java
@@ -1,6 +1,6 @@
-package com.lolduo.duo.repository;
+package com.lolduo.duo.repository.gameInfo;
 
-import com.lolduo.duo.entity.TrioEntity;
+import com.lolduo.duo.entity.gameInfo.TrioEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TrioRepository extends JpaRepository<TrioEntity,Long> {
